services/mirofish/backend/requirements.txt
# ===========================================
# MiroFish Backend Dependencies
# ===========================================
# Python 3.11+ required
# Install: pip install -r requirements.txt
# ===========================================

# ============= 核心框架 =============
flask>=3.0.0
flask-cors>=6.0.0

# ============= LLM 相关 =============
# OpenAI SDK（统一使用 OpenAI 格式调用 LLM）
openai>=1.0.0

# ============= Zep Cloud =============
zep-cloud==3.13.0

# ============= OASIS 社交媒体模拟 =============
# OASIS 社交模拟框架
camel-oasis==0.2.5
camel-ai==0.2.78

# ============= 文件处理 =============
PyMuPDF>=1.24.0
# 编码检测（支持非UTF-8编码的文本文件）
charset-normalizer>=3.0.0
chardet>=5.0.0

# ============= 工具库 =============
# 环境变量加载
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.0.0

# ============= Discord & Supabase =============
nextcord>=2.6.0
supabase>=2.5.0
pyjwt>=2.8.0
